# healthy-diet-api/openapi.yml
openapi: 3.0.3
info:
  title: Healthy Diet API
  version: 1.0.0
  description: API specification synchronized with current Axum routes.
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /consult:
    post:
      tags: [Nutrition]
      summary: Ask AI nutrition consult
      operationId: consult
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsultRequest'
      responses:
        '200':
          description: AI reply
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server/AI error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    ConsultRequest:
      type: object
      required: [question]
      properties:
        question:
          type: string
    ConsultResponse:
      type: object
      required: [reply]
      properties:
        reply:
          type: string
...